\name{OptionsSVM-package}
\alias{OptionsSVM-package}
\alias{OptionsSVM}
\docType{package}
\title{
\packageTitle{OptionsSVM}
}
\description{
\packageDescription{OptionsSVM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{OptionsSVM}
\packageIndices{OptionsSVM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{OptionsSVM}

Maintainer: \packageMaintainer{OptionsSVM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
